requirements.txt
Django==3.2.16
djangorestframework==3.12.4
python-dotenv==1.0.0
argon2-cffi==23.1.0
djangorestframework-simplejwt==4.7.2
Pillow==9.0.0
drf-spectacular==0.27.0
celery==5.4.0
redis==5.0.4
gunicorn==20.1.0
psycopg2-binary==2.9.9

